move key-store to java
